Swap the main Route class into Router and use Route as a Facade
Also change the NS to Router.
